refactor: AverageReview를 직접 반환하는 메서드로 변경 (#58)

+++ src/main/java/modelly/modelly_be/domain/recruitment/dto/response/GuestRecruitmentResponseDto.java
@@ -25,7 +25,7 @@ public record GuestRecruitmentResponseDto(
     boolean agreeInsta,
     boolean agreeMosaic,
     String etc,
-    int reviewCount,
+    Long reviewCount,
     double averageRating
 ) {
