+++ src/main/java/modelly/modelly_be/domain/review/dto/internal/AverageReview.java
@@ -4,7 +4,7 @@ import lombok.Builder;
 
 @Builder
 public record AverageReview(
-        Integer totalCount,
+        Long totalCount,
         Double averageRating
 ) {
 }
